+++ node/AudioProcessNode/node_spectrogram.py
@@ -99,7 +99,7 @@ class FactoryNode:
         use_pref_counter = node._opencv_setting_dict['use_pref_counter']
 
         # Create black texture for initial display
-        black_image = np.zeros((small_window_h, small_window_w, 3))
+        black_image = np.zeros((small_window_h, small_window_w, 3), dtype=np.uint8)
         black_texture = node.convert_cv_to_dpg(
             black_image,
             small_window_w,
