+++ node/ProcessNode/node_simple_filter.py
@@ -72,7 +72,7 @@ class FactoryNode:
         use_pref_counter = node._opencv_setting_dict['use_pref_counter']
 
 
-        black_image = np.zeros((small_window_w, small_window_h, 3))
+        black_image = np.zeros((small_window_w, small_window_h, 3), dtype=np.uint8)
         black_texture = convert_cv_to_dpg(
             black_image,
             small_window_w,
@@ -315,7 +315,7 @@ class Node(Node):
         use_pref_counter = self._opencv_setting_dict['use_pref_counter']
 
         # 初期化用黒画像
-        black_image = np.zeros((small_window_w, small_window_h, 3))
+        black_image = np.zeros((small_window_w, small_window_h, 3), dtype=np.uint8)
         black_texture = convert_cv_to_dpg(
             black_image,
             small_window_w,
